+++ CHANGELOG.md
@@ -19,10 +19,13 @@ DONE
     - After you turn it into a groupchat, you can start a new DM (with the first two ppl)
 
   - Leave Chat done
+  - Admin can remove user from chat
 
 TODO
 - Groupchat
-  - Make it look less ugly
+  - Make it look less ugly 
+  - Add "XYZ Left the group" message
+  - Move add User / remove User to a groupchat settings
 - OLD STUFF, NOT IMPORTANT YET
   - Chatbot.js:
     - Rate limit (so that quota for day isnt filled)